+++ exporters/otlp/otlplog/otlploggrpc/client.go
@@ -1,0 +1,14 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package otlploggrpc // import "go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc"
+
+type client struct {
+	// TODO: implement.
+}
+
+// newClient creates a new gRPC log client.
+func newClient(cfg config) (*client, error) {
+	// TODO: implement.
+	return &client{}, nil
+}